factor code to smaller crates
* async and tokio runtime are factored into a cli specific trait
* consts are moved to config
