+++ config/src/lib.rs
@@ -5,6 +5,15 @@ use zcash_primitives::{
 
 pub const DEFAULT_SERVER: &str = "https://lwdv3.zecwallet.co";
 pub const MAX_REORG: usize = 100;
+pub const WALLET_NAME: &str = "zingo-wallet.dat";
+pub const LOGFILE_NAME: &str = "zingo-wallet.debug.log";
+pub const ANCHOR_OFFSET: [u32; 5] = [4, 0, 0, 0, 0];
+pub const GAP_RULE_UNUSED_ADDRESSES: usize = if cfg!(any(target_os = "ios", target_os = "android"))
+{
+    0
+} else {
+    5
+};
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Network {
